# openapi/openapi.yaml
openapi: 3.1.1
info:
  title: oneSCM Handbook Studio API
  version: 0.12.0
  description: |
    Revisionssichere Erstellung rollen- und spartenspezifischer Benutzerhandbücher aus Markdown-Quellen.
    Jede Operation referenziert über `x-requirements` die zugehörigen User Stories (US-020).
    Operationen mit `x-extension: true` ergänzen die verbindliche Operationsliste aus Masterprompt §10 (siehe docs/01-gap-analyse.md, Lücke 10).
    Fehler werden als `application/problem+json` (RFC 9457) geliefert.
    Authentifizierung (ENTSCHEIDUNG E-15), je nach Betriebsart (`GET /auth/config`):
    - `oidc`: `Authorization: Bearer <JWT>` eines OpenID-Connect-Providers; technische Berechtigungen aus einem Gruppen-/Rollen-Claim.
    - `demo`: Header `X-User-Id` mit einem Demo-Benutzer; ohne Header nur Lesezugriff. Nur für Entwicklung und Demo.
    Projekt (ADR-014): Header `X-Project-Id` wählt das Projekt aller projektbezogenen Endpunkte (Standard `p_default`);
    IDs anderer Projekte werden wie nicht vorhandene behandelt (404). Ausnahmen: `/projects…` und die Health-Checks.
    Betrieb (ADR-015, ADR-027): jede Antwort trägt `X-Request-Id` (bei aktivem Tracing zusätzlich `X-Trace-Id`, W3C `traceparent` wird übernommen);
    Rate-Limits je Benutzer und Minute (429 mit `Retry-After`), mit `RATE_LIMIT_STORE=db` gemeinsam über alle Instanzen;
    Prometheus-Metriken unter `/metrics` (außerhalb von `/api/v1`, Bearer `METRICS_TOKEN` oder Berechtigung admin).
servers:
  - url: /api/v1
...
components:
  securitySchemes:
    oidcBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access-Token des konfigurierten OIDC-Providers (AUTH_MODE=oidc)
    demoUser:
      type: apiKey
      in: header
      name: X-User-Id
  parameters:
...
    OutlineId: { name: outlineId, in: path, required: true, schema: { type: string } }
    NodeId: { name: nodeId, in: path, required: true, schema: { type: string } }
...
  responses:
...
    Unauthorized:
      description: Nicht angemeldet oder Token ungültig
      headers:
        WWW-Authenticate: { schema: { type: string, example: Bearer } }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Problem:
      description: Fehler (RFC 9457)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        gate: { $ref: '#/components/schemas/GateResult' }
        blockers: { type: array, items: { type: object } }
...
    GateResult:
      type: object
      properties:
        passed: { type: boolean }
        checks:
          type: array
          items:
            type: object
            properties:
              code: { type: string }
              label: { type: string }
              passed: { type: boolean }
              details: { type: array, items: { type: string } }
...
security:
  - oidcBearer: []
  - demoUser: []
paths:
...
  /settings:
    get:
      operationId: getSettings
      summary: Einstellungen (Schwellenwerte, Uploadgrenzen, Terminologie)
      tags: [System]
      x-requirements: [US-005, US-001, US-015]
      x-extension: true
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: Einstellungen }
    put:
      operationId: putSettings
      summary: Einstellungen ändern (Berechtigung admin)
      tags: [System]
      x-requirements: [US-005, US-001, US-015]
      x-extension: true
      requestBody: { required: true, content: { application/json: { schema: { type: object } } } }
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: Einstellungen }
        '403': { $ref: '#/components/responses/Problem' }
...
  /media/{mediaSha}:
    get:
      operationId: getMedia
      summary: Bild ausliefern (nur eigenes Projekt, nosniff, unveränderlich zwischenspeicherbar)
      tags: [Medien]
      x-requirements: [NFR-16]
      x-extension: true
      parameters: [{ name: mediaSha, in: path, required: true, schema: { type: string, pattern: '^[a-f0-9]{64}$' } }]
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: 'Bilddaten, content: { image/png: {}, image/jpeg: {}, image/gif: {}, image/webp: {} }' }
        '404': { $ref: '#/components/responses/Problem' }
    patch:
      operationId: setMediaTitle
      summary: Bildtitel für das Bildverzeichnis setzen (Berechtigung edit)
      tags: [Stammdaten]
      x-requirements: [NFR-19, NFR-16]
      x-extension: true
      parameters: [{ name: mediaSha, in: path, required: true, schema: { type: string } }]
      requestBody:
        content: { application/json: { schema: { type: object, properties: { title: { type: [string, 'null'] } } } } }
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: OK }
...
  /outlines:
    get:
      operationId: listOutlines
      summary: 'Gliederungen des Projekts mit Variante, Version, Anzahl Einträge/Schnipsel und Projektmärkten (ADR-032)'
      tags: [Stammdaten]
      x-requirements: [NFR-19]
      x-extension: true
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: 'OK' }
    post:
      operationId: createOutline
      summary: 'Gliederung anlegen: leer, aus der Kapitelstruktur (fromChapters), aus nodes oder per Upload (content + format markdown|json); Berechtigung edit'
      tags: [Stammdaten]
      x-requirements: [NFR-19]
      x-extension: true
      requestBody:
        content: { application/json: { schema: { type: object, properties: { name: { type: string }, roles: { type: array, items: { type: string } }, divisions: { type: array, items: { type: string } }, marketScope: { type: string, enum: [blueprint, markets] }, markets: { type: array, items: { type: string } }, fromChapters: { type: boolean }, content: { type: string }, format: { type: string, enum: [markdown, json] }, nodes: { type: array, items: { type: object } } } } } }
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '201': { description: 'Angelegt' }
        '400': { description: 'Ungültige Variante oder keine Gliederungseinträge' }
  /outlines/{outlineId}:
    get:
      operationId: getOutline
      summary: 'Gliederung mit nummerierten Einträgen und Versionen'
      tags: [Stammdaten]
      x-requirements: [NFR-19]
      x-extension: true
      parameters: [{ $ref: '#/components/parameters/OutlineId' }]
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: 'OK' }
        '404': { description: 'Unbekannt' }
    patch:
      operationId: updateOutline
      summary: 'Name, Variante oder Status ändern (höchstens eine aktive Version je Gliederung)'
      tags: [Stammdaten]
      x-requirements: [NFR-19]
      x-extension: true
      parameters: [{ $ref: '#/components/parameters/OutlineId' }]
      requestBody:
        content: { application/json: { schema: { type: object, properties: { name: { type: string }, description: { type: [string, 'null'] }, roles: { type: array, items: { type: string } }, divisions: { type: array, items: { type: string } }, marketScope: { type: string, enum: [blueprint, markets] }, markets: { type: array, items: { type: string } }, status: { type: string, enum: [draft, active, archived] } } } } }
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: 'OK' }
        '400': { description: 'Ungültig' }
    delete:
      operationId: deleteOutline
      summary: 'Version löschen (nicht, wenn weitere Versionen darauf aufbauen)'
      tags: [Stammdaten]
      x-requirements: [NFR-19]
      x-extension: true
      parameters: [{ $ref: '#/components/parameters/OutlineId' }]
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '204': { description: 'Gelöscht' }
        '409': { description: 'Grundlage weiterer Versionen' }
  /outlines/{outlineId}/versions:
    post:
      operationId: createOutlineVersion
      summary: 'Als neue Version speichern (Kopie mit Zuordnungen und Planung)'
      tags: [Stammdaten]
      x-requirements: [NFR-19]
      x-extension: true
      parameters: [{ $ref: '#/components/parameters/OutlineId' }]
      requestBody:
        content: { application/json: { schema: { type: object, properties: { name: { type: string } } } } }
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '201': { description: 'Neue Version' }
  /outlines/{outlineId}/export:
    get:
      operationId: exportOutline
      summary: 'Gliederung als Markdown oder JSON (Format für den Upload)'
      tags: [Stammdaten]
      x-requirements: [NFR-19]
      x-extension: true
      parameters: [{ $ref: '#/components/parameters/OutlineId' }, { name: format, in: query, schema: { type: string, enum: [md, json] } }]
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: 'OK' }
  /outlines/{outlineId}/nodes:
    post:
      operationId: addOutlineNode
      summary: 'Kapitel oder Unterkapitel (parentId) hinzufügen, optional nach afterId'
      tags: [Stammdaten]
      x-requirements: [NFR-19]
      x-extension: true
      parameters: [{ $ref: '#/components/parameters/OutlineId' }]
      requestBody:
        content: { application/json: { schema: { type: object, required: [title], properties: { title: { type: string }, parentId: { type: [string, 'null'] }, afterId: { type: [string, 'null'] }, description: { type: string } } } } }
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '201': { description: 'Hinzugefügt' }
        '400': { description: 'Ungültig' }
  /outline-nodes/{nodeId}:
    patch:
      operationId: updateOutlineNode
      summary: 'Eintrag umbenennen, verschieben (move up|down), einordnen (parentId; null = Kapitel) oder per Drag & Drop vor einen Eintrag setzen (beforeId, übernimmt dessen Ebene)'
      tags: [Stammdaten]
      x-requirements: [NFR-19, NFR-22]
      x-extension: true
      parameters: [{ $ref: '#/components/parameters/NodeId' }]
      requestBody:
        content: { application/json: { schema: { type: object, properties: { title: { type: string }, description: { type: [string, 'null'] }, move: { type: string, enum: [up, down] }, parentId: { type: [string, 'null'] }, beforeId: { type: [string, 'null'] } } } } }
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: 'OK' }
    delete:
      operationId: deleteOutlineNode
      summary: 'Eintrag samt Unterkapiteln löschen; Zuordnungen und Planung entfallen'
      tags: [Stammdaten]
      x-requirements: [NFR-19]
      x-extension: true
      parameters: [{ $ref: '#/components/parameters/NodeId' }]
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: 'OK' }
  /outline-nodes/{nodeId}/plan:
    put:
      operationId: setPlanItem
      summary: 'Redaktionsplanung eines Eintrags: Verantwortliche, Termin (JJJJ-MM-TT), Status, Notiz'
      tags: [Stammdaten]
      x-requirements: [NFR-19]
      x-extension: true
      parameters: [{ $ref: '#/components/parameters/NodeId' }]
      requestBody:
        content: { application/json: { schema: { type: object, properties: { assignee: { type: [string, 'null'] }, dueDate: { type: [string, 'null'], format: date }, status: { type: string, enum: [open, in_progress, review, done] }, note: { type: [string, 'null'] } } } } }
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: 'OK' }
        '400': { description: 'Ungültig' }
  /outlines/{outlineId}/plan:
    get:
      operationId: getOutlinePlan
      summary: 'Redaktionsplan einer Gliederung mit Überfälligkeit und Fortschritt'
      tags: [Stammdaten]
      x-requirements: [NFR-19]
      x-extension: true
      parameters: [{ $ref: '#/components/parameters/OutlineId' }]
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: 'OK' }
  /outlines/{outlineId}/draft:
    get:
      operationId: getDraftManual
      summary: 'Draft Manual: Einträge mit zugeordneten Schnipseln und Kennzeichnung (duplicate, contradiction, gap, warning) samt Übersicht (ADR-033)'
      tags: [Draft Manual]
      x-requirements: [NFR-20]
      x-extension: true
      parameters: [{ $ref: '#/components/parameters/OutlineId' }]
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: 'OK' }
  /outlines/{outlineId}/draft/export:
    get:
      operationId: exportDraftManual
      summary: 'Draft Manual als Markdown (Kennzeichnungen als Zitatzeilen; flags=false ohne)'
      tags: [Draft Manual]
      x-requirements: [NFR-20]
      x-extension: true
      parameters: [{ $ref: '#/components/parameters/OutlineId' }, { name: flags, in: query, schema: { type: boolean } }]
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: 'OK' }
  /outlines/{outlineId}/candidates:
    get:
      operationId: listOutlineCandidates
      summary: 'Nicht zugeordnete aktuelle Schnipsel (Suche q, chapterId, Seiten; assigned=all für alle) mit Kennzeichnung'
      tags: [Draft Manual]
      x-requirements: [NFR-20]
      x-extension: true
      parameters: [{ $ref: '#/components/parameters/OutlineId' }, { name: q, in: query, schema: { type: string } }, { name: chapterId, in: query, schema: { type: string } }, { name: page, in: query, schema: { type: integer } }, { name: pageSize, in: query, schema: { type: integer, maximum: 200 } }]
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: 'OK' }
  /outlines/{outlineId}/assignments:
    post:
      operationId: assignOutlineSnippets
      summary: 'Schnipsel einem Eintrag zuordnen (bereits zugeordnete werden verschoben; höchstens 500); beforeSnippetId: davor einfügen (Drag & Drop)'
      tags: [Draft Manual]
      x-requirements: [NFR-20, NFR-22]
      x-extension: true
      parameters: [{ $ref: '#/components/parameters/OutlineId' }]
      requestBody:
        content: { application/json: { schema: { type: object, required: [nodeId, snippetIds], properties: { nodeId: { type: string }, snippetIds: { type: array, items: { type: string } }, beforeSnippetId: { type: string } } } } }
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: 'OK' }
        '400': { description: 'Ungültig' }
  /outlines/{outlineId}/assignments/{snippetId}:
    patch:
      operationId: moveOutlineAssignment
      summary: 'Schnipsel innerhalb des Eintrags verschieben'
      tags: [Draft Manual]
      x-requirements: [NFR-20]
      x-extension: true
      parameters: [{ $ref: '#/components/parameters/OutlineId' }, { name: snippetId, in: path, required: true, schema: { type: string } }]
      requestBody:
        content: { application/json: { schema: { type: object, required: [move], properties: { move: { type: string, enum: [up, down] } } } } }
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '204': { description: 'Verschoben' }
    delete:
      operationId: unassignOutlineSnippet
      summary: 'Zuordnung lösen'
      tags: [Draft Manual]
      x-requirements: [NFR-20]
      x-extension: true
      parameters: [{ $ref: '#/components/parameters/OutlineId' }, { name: snippetId, in: path, required: true, schema: { type: string } }]
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '204': { description: 'Gelöst' }
        '404': { description: 'Nicht zugeordnet' }
  /outlines/{outlineId}/auto-assign:
    post:
      operationId: autoAssignOutline
      summary: 'Automatisch zuordnen über Kapitel-/Unterkapiteltitel der Quellen'
      tags: [Draft Manual]
      x-requirements: [NFR-20]
      x-extension: true
      parameters: [{ $ref: '#/components/parameters/OutlineId' }]
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: '{ assigned, remaining }' }
  /outlines/{outlineId}/compare:
    get:
      operationId: compareOutlineVersions
      summary: 'Zwei Versionen derselben Gliederung vergleichen (Einträge über stabile Kennung, Titel, Position, Ebene, Zuordnungen, Variante; ADR-035)'
      tags: [Stammdaten]
      x-requirements: [NFR-22]
      x-extension: true
      parameters:
        - { $ref: '#/components/parameters/OutlineId' }
        - { name: with, in: query, required: true, schema: { type: string }, description: 'Kennung der Vergleichsversion (ältere Seite)' }
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: '{ from, to, summary, variant, entries }' }
        '400': { description: 'andere Gliederung oder with fehlt' }
  /outlines/{outlineId}/chapters:
    get:
      operationId: listVariantChapters
      summary: 'Kapitel der Handbuch-Variante mit Versions- und Freigabestand (ADR-034)'
      tags: [Draft Manual]
      x-requirements: [NFR-21]
      x-extension: true
      parameters: [{ $ref: '#/components/parameters/OutlineId' }]
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: '{ outlineId, familyId, name, chapters }' }
  /outlines/{outlineId}/materialize:
    post:
      operationId: materializeVariant
      summary: 'Kapitel der Variante anlegen bzw. aktualisieren (je Kapitel der Gliederung, fortgeführt über Gliederungsversionen)'
      tags: [Draft Manual]
      x-requirements: [NFR-21]
      x-extension: true
      parameters: [{ $ref: '#/components/parameters/OutlineId' }]
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: 'Kapitel der Variante' }
        '400': { description: 'Gliederung ohne Kapitel' }
  /outlines/{outlineId}/generate:
    post:
      operationId: generateVariant
      summary: 'Kapitelentwürfe der Variante aus den Zuordnungen erzeugen; danach Werkstatt, Qualitätsgate und Freigabe wie bei Quellenkapiteln'
      tags: [Draft Manual]
      x-requirements: [NFR-21]
      x-extension: true
      parameters: [{ $ref: '#/components/parameters/OutlineId' }]
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: '{ outlineId, generated, results: [{ chapterId, title, status: generated|skipped|blocked, versionNo, message }] }' }
...
  /master-data/import:
    post:
      operationId: importMasterData
      summary: 'Abkürzungen, Glossar oder FAQ aus CSV (; , Tab) oder Excel (.xlsx) importieren – Vorschau oder Übernahme (ADR-036)'
      tags: [Stammdaten]
      x-requirements: [NFR-23]
      x-extension: true
      parameters:
        - { name: kind, in: query, required: true, schema: { type: string, enum: [abbreviations, glossary, faq] } }
        - { name: apply, in: query, schema: { type: boolean, default: false } }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { type: object, required: [file], properties: { file: { type: string, format: binary } } }
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: '{ kind, applied, summary: { rows, create, update, unchanged, error }, rows: [{ row, key, action, message }] }' }
        '400': { description: 'Spalten fehlen, leere Tabelle oder unbekannte Art' }
        '415': { description: 'nur .csv und .xlsx' }
  /abbreviations:
    get:
      operationId: listAbbreviations
      summary: 'Abkürzungsverzeichnis'
      tags: [Stammdaten]
      x-requirements: [NFR-19]
      x-extension: true
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: 'OK' }
    post:
      operationId: createAbbreviation
      summary: 'Abkürzung erfassen (je Projekt eindeutig; Berechtigung edit)'
      tags: [Stammdaten]
      x-requirements: [NFR-19]
      x-extension: true
      requestBody:
        content: { application/json: { schema: { type: object, required: [abbreviation, expansion], properties: { abbreviation: { type: string }, expansion: { type: string }, description: { type: string } } } } }
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '201': { description: 'Angelegt' }
        '409': { description: 'Existiert bereits' }
  /abbreviations/suggestions:
    get:
      operationId: listAbbreviationSuggestions
      summary: 'In aktuellen Schnipseln vorkommende, noch nicht erfasste Abkürzungen'
      tags: [Stammdaten]
      x-requirements: [NFR-19]
      x-extension: true
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: 'OK' }
  /abbreviations/{abbreviationId}:
    patch:
      operationId: updateAbbreviation
      summary: 'Abkürzung ändern'
      tags: [Stammdaten]
      x-requirements: [NFR-19]
      x-extension: true
      parameters: [{ name: abbreviationId, in: path, required: true, schema: { type: string } }]
      requestBody:
        content: { application/json: { schema: { type: object } } }
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: 'OK' }
    delete:
      operationId: deleteAbbreviation
      summary: 'Abkürzung löschen'
      tags: [Stammdaten]
      x-requirements: [NFR-19]
      x-extension: true
      parameters: [{ name: abbreviationId, in: path, required: true, schema: { type: string } }]
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '204': { description: 'Gelöscht' }
  /faq:
    get:
      operationId: listFaq
      summary: 'FAQ-Einträge, gefiltert nach role, division, status, language'
      tags: [Stammdaten]
      x-requirements: [NFR-19]
      x-extension: true
      parameters: [{ name: role, in: query, schema: { type: string } }, { name: division, in: query, schema: { type: string } }, { name: status, in: query, schema: { type: string, enum: [draft, published] } }]
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: 'OK' }
    post:
      operationId: createFaq
      summary: 'FAQ-Eintrag anlegen (sourceQuestion kennzeichnet einen übernommenen Vorschlag)'
      tags: [Stammdaten]
      x-requirements: [NFR-19]
      x-extension: true
      requestBody:
        content: { application/json: { schema: { type: object, required: [question, answer], properties: { question: { type: string }, answer: { type: string }, roles: { type: array, items: { type: string } }, divisions: { type: array, items: { type: string } }, status: { type: string }, sourceQuestion: { type: string } } } } }
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '201': { description: 'Angelegt' }
        '400': { description: 'Ungültig' }
  /faq/suggestions:
    get:
      operationId: listFaqSuggestions
      summary: 'Vorschläge aus dem Handbuch-Assistenten: wiederholte, unbeantwortete oder negativ bewertete Fragen (Berechtigung edit)'
      tags: [Stammdaten]
      x-requirements: [NFR-19, NFR-13]
      x-extension: true
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: 'OK' }
  /faq/{faqId}:
    patch:
      operationId: updateFaq
      summary: 'FAQ-Eintrag ändern oder verschieben (move up|down)'
      tags: [Stammdaten]
      x-requirements: [NFR-19]
      x-extension: true
      parameters: [{ name: faqId, in: path, required: true, schema: { type: string } }]
      requestBody:
        content: { application/json: { schema: { type: object } } }
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: 'OK' }
    delete:
      operationId: deleteFaq
      summary: 'FAQ-Eintrag löschen'
      tags: [Stammdaten]
      x-requirements: [NFR-19]
      x-extension: true
      parameters: [{ name: faqId, in: path, required: true, schema: { type: string } }]
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '204': { description: 'Gelöscht' }
  /image-index:
    get:
      operationId: getImageIndex
      summary: 'Bildverzeichnis: Abbildungsnummer, Titel, Alternativtexte, Verwendung in Quellen und Kapitelversionen'
      tags: [Stammdaten]
      x-requirements: [NFR-19, NFR-16]
      x-extension: true
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: 'OK' }